update: added more protected routes
