+++ server/api/guests.js
@@ -1,12 +1,20 @@
 const router = require('express').Router()
+const jwt = require('jsonwebtoken')
+const verifyToken = require('./verifyToken')
 const {Guest} = require('../db/models')
 module.exports = router
 
-router.post('/', async (req, res, next) => {
+//route protection- successful!
+router.post('/', verifyToken, async (req, res, next) => {
   console.log('**POST api/guests**')
   try {
-    const guest = await Guest.create(req.body)
-    res.status(201).send(guest)
+    const decoded = jwt.verify(req.token, 'secretkey')
+    if (!decoded) {
+      res.send(403)
+    } else {
+      const guest = await Guest.create(req.body)
+      res.status(201).send(guest)
+    }
   } catch (error) {
     next(error)
   }
